requirements.txt
strands-agents[gemini]
strands-agents-tools
a2a-sdk
python-dotenv
fastapi
uvicorn
httpx
pytest
